refactor: quantization for embedding layers
